requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
Django==4.1.4
django-environ==0.9.0
django-graphql-jwt==0.3.4
djangorestframework==3.14.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
Pillow==9.5.0
promise==2.3
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
Unidecode==1.3.6
